refactor: split out common dataset logic into `BaseDataset`
